requirements.txt
python-dotenv~=1.1.1
openai~=1.107.0
python-telegram-bot~=22.3
pydub~=0.25.1
requests~=2.32.5
aiogram~=3.22.0